Add corporate integration tests
