+++ src/itest/java/uk/gov/companieshouse/disqualifiedofficersdataapi/steps/CommonApiSteps.java
@@ -1,11 +1,8 @@
 package uk.gov.companieshouse.disqualifiedofficersdataapi.steps;
 
 import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -21,16 +18,6 @@ public class CommonApiSteps {
     public void theApplicationRunning() {
         assertThat(restTemplate).isNotNull();
         lastResponse = null;
-    }
-
-    @When("the client invokes {string} endpoint")
-    public void theClientInvokesAnEndpoint(String url) {
-        lastResponse = restTemplate.getForEntity(url, String.class);
-    }
-
-    @Then("the client receives status code of {int}")
-    public void theClientReceivesStatusCodeOf(int code) {
-        assertThat(lastResponse.getStatusCode()).isEqualTo(HttpStatus.valueOf(code));
     }
 
 }
