--- src/itest/resources/features/Heath_Check_API.feature
@@ -1,8 +1,0 @@
-Feature: Health check API endpoint
-
-  Scenario: Client invokes GET /healthcheck endpoint
-    Given disqualified officers data api service is running
-
-    When the client invokes '/healthcheck' endpoint
-
-    Then the client receives status code of 200
